Fix standalone tests.
